backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
langchain
langchain-community
langchain-google-genai
notion-client
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
pgvector
alembic
